Add files via upload
This is a vulnerability assessment report demonstration where a server hosting a database was assessed for weaknesses that posed risks to the organization.  In this demonstration, I site the purpose, risk and remediation strategy to help aid in strengthening the security posture of the organization. The NIST frameworks that were considered included NIST SP 800-30 Rev 1 for risk assessment best practices, NIST SP 800-53 Rev 5 for security control advisory and NIST SP 800-171 Rev 3 for securing a database, specifically.